+++ app/resources/subtask.py
@@ -86,6 +86,10 @@ class SubTask(Resource):
 
     @f_jwt.jwt_required()
     def put(self, task_id, subtask_id):
+        user_id = f_jwt.get_jwt_identity()
+        # user_id=20
+        app.logger.debug("user_id= %s", user_id)
+
         data = request.get_json()
         subtask_dict = json.loads(json.dumps(data))
         current_time = datetime.now()
@@ -94,7 +98,7 @@ class SubTask(Resource):
         UPDATE_SUBTASK = '''UPDATE subtasks SET title= %s, description= %s, status= %s,
         plan_start_date= %s, plan_end_date= %s, actual_end_date= %s, duration= %s, 
         task_type= %s, notify= %s, repeat= %s, priority= %s, updated_at= %s 
-        WHERE id= %s AND task_id= %s'''
+        WHERE id= %s AND task_id= %s AND user_id= %s'''
 
         # catch exception for invalid SQL statement
         try:
@@ -105,7 +109,10 @@ class SubTask(Resource):
             cursor.execute(UPDATE_SUBTASK, (subtask_dict['title'], subtask_dict['description'], subtask_dict['status'],
                             subtask_dict['plan_start_date'], subtask_dict['plan_end_date'], subtask_dict['actual_end_date'],
                             subtask_dict['duration'], subtask_dict['task_type'], subtask_dict['notify'], subtask_dict['repeat'],
-                            subtask_dict['priority'], current_time, subtask_id, task_id,))
+                            subtask_dict['priority'], current_time, subtask_id, task_id, user_id,))
+            # app.logger.debug("row_counts= %s", cursor.rowcount)
+            if cursor.rowcount != 1:
+                abort(400, 'Bad Request: update row error')
         except (Exception, psycopg2.Error) as err:
             app.logger.debug(err)
             abort(400, 'Bad Request')
@@ -115,7 +122,11 @@ class SubTask(Resource):
 
     @f_jwt.jwt_required()
     def delete(self, task_id, subtask_id):
-        DELETE_SUBTASK = 'DELETE FROM subtasks WHERE id= %s AND task_id= %s'
+        user_id = f_jwt.get_jwt_identity()
+        # user_id=20
+        app.logger.debug("user_id= %s", user_id)
+
+        DELETE_SUBTASK = 'DELETE FROM subtasks WHERE id= %s AND task_id= %s AND user_id= %s'
 
         # catch exception for invalid SQL statement
         try:
@@ -123,7 +134,7 @@ class SubTask(Resource):
             cursor = main.db_conn.cursor()
             # app.logger.debug("cursor object: %s", cursor)
 
-            cursor.execute(DELETE_SUBTASK, (subtask_id, task_id,))
+            cursor.execute(DELETE_SUBTASK, (subtask_id, task_id, user_id,))
         except (Exception, psycopg2.Error) as err:
             app.logger.debug(err)
             abort(400, 'Bad Request')
